Caption Refactor: Extract text field and highlighting. Caption Generate: Load existing caption from cache or file when generation ends while another image is selected. Caption Groups: Show menu with expanded wildcards when clicking a button that uses wildcards.
