+++ caption/caption_generate.py
@@ -1,4 +1,5 @@
 import os, re, traceback
+from enum import Enum
 from PySide6 import QtWidgets
 from PySide6.QtCore import Qt, Slot, Signal, QRunnable, QObject, QSignalBlocker
 from infer import Inference, InferencePresetWidget, TagPresetWidget, PromptWidget, PromptsHighlighter, InferenceProcess
@@ -7,6 +8,12 @@ from lib.filelist import DataKeys
 import lib.qtlib as qtlib
 from config import Config
 from .caption_tab import CaptionTab
+
+
+class ApplyMode(Enum):
+    Append  = "append"
+    Prepend = "prepend"
+    Replace = "replace"
 
 
 class CaptionGenerate(CaptionTab):
@@ -71,9 +78,9 @@ class CaptionGenerate(CaptionTab):
         layout.addWidget(self.tagSettings, row, 0, 2, 2)
 
         self.cboMode = QtWidgets.QComboBox()
-        self.cboMode.addItem("Append")
-        self.cboMode.addItem("Prepend")
-        self.cboMode.addItem("Replace")
+        self.cboMode.addItem("Append", ApplyMode.Append)
+        self.cboMode.addItem("Prepend", ApplyMode.Prepend)
+        self.cboMode.addItem("Replace", ApplyMode.Replace)
         layout.addWidget(self.cboMode, row, 3)
 
         self.cboCapTag = QtWidgets.QComboBox()
@@ -120,7 +127,7 @@ class CaptionGenerate(CaptionTab):
 
     def updatePreview(self, prompt: str):
         if self._hasRefinedVar:
-            self._parser.updateRefinedCaption(self.ctx.container.getCaption())
+            self._parser.updateRefinedCaption(self.ctx.text.getCaption())
 
         preview, varPositions = self._parser.parseWithPositions(prompt)
         self.txtPromptPreview.setPlainText(preview)
@@ -176,24 +183,42 @@ class CaptionGenerate(CaptionTab):
         self.statusBar.showMessage(message)
 
     @Slot()
-    def onGenerated(self, imgPath, text):
+    def onGenerated(self, imgPath: str, generatedText: str):
         self.btnGenerate.setEnabled(True)
 
-        if not text:
+        if not generatedText:
             self.statusBar.showColoredMessage("Finished with empty result", False, 0)
             return
+
         self.statusBar.showColoredMessage("Done", True)
 
         filelist = self.ctx.tab.filelist
         if imgPath == filelist.getCurrentFile():
-            mode = self.cboMode.currentText()
-            self.ctx.captionGenerated.emit(text, mode)
+            text = self._addToCaption(generatedText, self.ctx.text.getCaption())
+            self.ctx.text.setCaption(text)
+            self.ctx.needsRulesApplied.emit()
         else:
+            existingText = filelist.getData(imgPath, DataKeys.Caption)
+            if existingText is None:
+                existingText = self.ctx.container.srcSelector.loadCaption(imgPath)
+
+            text = self._addToCaption(generatedText, existingText)
             filelist.setData(imgPath, DataKeys.Caption, text)
             filelist.setData(imgPath, DataKeys.CaptionState, DataKeys.IconStates.Changed)
 
+    def _addToCaption(self, generatedText: str, existingText: str | None) -> str:
+        if not existingText:
+            return generatedText
+
+        match self.cboMode.currentData():
+            case ApplyMode.Append:  return f"{existingText}{os.linesep}{generatedText}"
+            case ApplyMode.Prepend: return f"{generatedText}{os.linesep}{existingText}"
+            case ApplyMode.Replace: return generatedText
+            case _:
+                raise ValueError("Invalid mode")
+
     @Slot()
-    def onFail(self, errorMsg):
+    def onFail(self, errorMsg: str):
         self.btnGenerate.setEnabled(True)
         self.statusBar.showColoredMessage(errorMsg, False, 0)
 
@@ -220,7 +245,7 @@ class CurrentVariableParser(TemplateVariableParser):
     def _getImgProperties(self, var: str) -> str | None:
         match var:
             case self.CURRENT_VAR_NAME:
-                return self.ctx.container.getCaption()
+                return self.ctx.text.getCaption()
             case self.REFINED_VAR_NAME:
                 return self.refinedCaption
